deleted eventForm store, added event validations
- the eventForm store was deleted since it can cause problems
- added some validations before creating and updating

+++ src/app/api/events/[id]/route.ts
@@ -1,5 +1,6 @@
 import { ERRORS } from '@/constants/backend/errors'
 import { prisma } from '@/lib/prisma'
+import { validateEventBody } from '@/lib/validateEvent'
 import { NextRequest, NextResponse } from 'next/server'
 
 // update an event
@@ -7,8 +8,24 @@ export async function PATCH(
   req: NextRequest,
   { params }: { params: { id: string } }
 ) {
-  const id = parseInt(params.id)
+  const id = Number(params.id)
   const body = await req.json()
+
+  if (isNaN(id)) {
+    return NextResponse.json(
+      { error: ERRORS.EVENT_BY_ID.INVALID_ID },
+      { status: 400 }
+    )
+  }
+
+  const error = validateEventBody(body)
+  if (error) return NextResponse.json({ error: error.error }, { status: error.status })
+
+  const existingEvent = await prisma.event.findUnique({ where: { id } })
+  if (!existingEvent) {
+    return NextResponse.json({ error: ERRORS.EVENT_BY_ID.NOT_FOUND }, { status: 404 })
+
+  }
 
   try {
     const updatedEvent = await prisma.event.update({
@@ -20,12 +37,16 @@ export async function PATCH(
         date: new Date(body.date),
         ...(body.images ? { images: { set: body.images } } : {}),
       },
-    })
+    }) 
 
     return NextResponse.json(updatedEvent)
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (error) {
-    return new NextResponse(ERRORS.EVENT_BY_ID.UPDATING, { status: 500 })
+    console.error('Error actualizando evento:', error)
+    return NextResponse.json(
+      { error: ERRORS.EVENT_BY_ID.UPDATING },
+      { status: 500 }
+    )
   }
 }
 
@@ -36,6 +57,19 @@ export async function DELETE(
 ) {
   const id = Number(params.id)
 
+  if (isNaN(id)) {
+    return NextResponse.json(
+      { error: ERRORS.EVENT_BY_ID.INVALID_ID },
+      { status: 400 }
+    )
+  }
+
+  const existingEvent = await prisma.event.findUnique({ where: { id } })
+  if (!existingEvent) {
+    return NextResponse.json({ error: ERRORS.EVENT_BY_ID.NOT_FOUND }, { status: 404 })
+
+  }
+
   try {
     const deletedEvent = await prisma.event.delete({
       where: { id },
@@ -44,7 +78,10 @@ export async function DELETE(
     return NextResponse.json(deletedEvent)
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (error) {
-    return new NextResponse(ERRORS.EVENT_BY_ID.DELETING, { status: 500 })
+    return NextResponse.json(
+      { error: ERRORS.EVENT_BY_ID.UPDATING },
+      { status: 500 }
+    )
   }
 }
 
@@ -53,18 +90,24 @@ export async function GET(
   request: Request,
   { params }: { params: { id: string } }
 ) {
-  const eventId = Number(params.id)
+  const id = Number(params.id)
 
-  if (isNaN(eventId)) {
+  if (isNaN(id)) {
     return NextResponse.json(
       { error: ERRORS.EVENT_BY_ID.INVALID_ID },
       { status: 400 }
     )
   }
 
+  const existingEvent = await prisma.event.findUnique({ where: { id } })
+  if (!existingEvent) {
+    return NextResponse.json({ error: ERRORS.EVENT_BY_ID.NOT_FOUND }, { status: 404 })
+
+  }
+
   try {
     const event = await prisma.event.findUnique({
-      where: { id: eventId },
+      where: { id: id },
       include: {
         creator: true,
         payments: true,
